fix crash on exit

+++ ST_engine/ST_message_bus/src/main/message.hpp
@@ -25,12 +25,12 @@ private:
     friend class message_allocator;
     std::shared_ptr<void> data; //yes, this holds anything created with make_data<>() AND calls the correct destructor
     //that's how shared_ptr works, if you don't believe me, well google it or something
-    uint16_t id; //used during allocation and deallocation
+    uint8_t id; //used during allocation and deallocation
 
-    message(int name, const std::shared_ptr<void>& data, uint16_t id);
+    message(int name, const std::shared_ptr<void>& data, uint8_t id);
 public:
     int32_t msg_name;
-    uint16_t get_id() const;
+    uint8_t get_id() const;
     void* get_data() const;
     message* make_copy() const;
 };
@@ -42,7 +42,7 @@ public:
 /**
  * constructor - only ever used by the message_allocator class.
  */
-inline message::message(int name, const std::shared_ptr<void>& data, uint16_t id){
+inline message::message(int name, const std::shared_ptr<void>& data, uint8_t id){
     msg_name = name;
     this->data = data;
     this->id = id;
@@ -58,7 +58,7 @@ inline void* message::get_data() const{
 /**
  * @return The id of the message.
  */
-inline uint16_t message::get_id() const{
+inline uint8_t message::get_id() const{
     return id;
 }
